Move diagnostics to the driver factory
